Board Initialization, Display, and Input
The Game Board is Displayed with Adjacent Mine Counts on Known Boxes.
The User Inputs whether there is a Mine or Not.
The Board Displays * as Marked Mine's Box, while # as a Safe Box.

First, the Mines are Placed Randomly on the Board.
Then on the Remaining Boxes, the Adjacent Mines are Counted.
Then, some known Boxes are Randomly made Unknown, and the Mines are Hidden, for the Puzzle.

The User Input and its Basic Filteration.