DetonateMines, Code Update, and ReadMe
Enter -8-8 as Input, to Detonate the Mines.
If Marked Correctly, You will Win, otherwise Lose.
Then Exact Mines Location will be Shown.

You can then PlayAgain with a New Baord.
With a New Game, the Baord first Needs to be Reset.
And then Board Initialization is Done.

The Board Initialization Code is Minimized.
The User Input is Filtered to Prevent Crash from Character Input.

The ReadMe File is Defined. The Back Story and Rules of Games.